fixup: restore ConversationLeftPanel and ConversationDetailPanel
Both wrappers were already thin aliases over scraps Flex, so removing
them in 3845762db6e was scope creep for a styled-component migration.
Restore the exports and put all three call sites back on the wrappers.
