+++ static/app/views/performance/newTraceDetails/traceDrawer/tabs/traceAiConversations.tsx
@@ -10,6 +10,8 @@ import {t} from 'sentry/locale';
 import {normalizeUrl} from 'sentry/utils/url/normalizeUrl';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {
+  ConversationDetailPanel,
+  ConversationLeftPanel,
   ConversationSplitLayout,
   ConversationViewSkeleton,
 } from 'sentry/views/explore/conversations/components/conversationLayout';
@@ -176,7 +178,6 @@ function TraceConversationChat({
   onSelectSpan: (spanId: string) => void;
   selectedSpanId: string | null;
 }) {
-  const organization = useOrganization();
   const {selectedNode, handleSelectNode} = useConversationSelection({
     nodes,
     selectedSpanId,
@@ -205,7 +206,7 @@ function TraceConversationChat({
       <ConversationSplitLayout
         sizeStorageKey="trace-conversation-split-size"
         left={
-          <Flex direction="column" flex={1} minHeight="0" overflow="hidden">
+          <ConversationLeftPanel>
             <Flex flex="1" minHeight="0" width="100%" overflowX="hidden" overflowY="auto">
               <MessagesPanel
                 nodes={nodes}
@@ -213,29 +214,13 @@ function TraceConversationChat({
                 onSelectNode={handleSelectNode}
               />
             </Flex>
-          </Flex>
+          </ConversationLeftPanel>
         }
         right={
-          <Flex
-            direction="column"
-            flex={1}
-            minHeight="0"
-            background="primary"
-            overflowY="auto"
-            overflowX="hidden"
-          >
-            {selectedNode?.renderDetails({
-              node: selectedNode,
-              manager: null,
-              onParentClick: () => {},
-              onTabScrollToNode: () => {},
-              organization,
-              replay: null,
-              traceId: nodeTraceMap.get(selectedNode.id) ?? '',
-              hideNodeActions: true,
-              initiallyCollapseAiIO: true,
-            })}
-          </Flex>
+          <ConversationDetailPanel
+            selectedNode={selectedNode}
+            nodeTraceMap={nodeTraceMap}
+          />
         }
       />
     </TraceStateProvider>
